create document

+++ src/Controller/User/DeleteUserController.php
@@ -4,6 +4,7 @@ namespace App\Controller\User;
 
 use App\Entity\User;
 use App\Repository\UserRepository;
+use App\Security\Voter\UserVoter;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -12,11 +13,13 @@ use Symfony\Component\Routing\Attribute\Route;
 class DeleteUserController extends AbstractController
 {
     #[Route('/user/remove/{id}', name: 'app_user_remove')]
-    public function remove(
+    public function remove(        
         User $user,
         UserRepository $repository,
         Request $request
-    ): Response {
+    ): Response {        
+        $this->denyAccessUnlessGranted(UserVoter::DELETE, $user);
+        
         $repository->remove($user);
 
         $this->addFlash('sucess', sprintf('Profesional eliminado'));
